% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accumulate_evidence_ddm}
\alias{accumulate_evidence_ddm}
\title{Simulate evidence accumulation in a drift-diffusion model}
\usage{
accumulate_evidence_ddm(
  A,
  V,
  Z,
  ndt,
  max_t,
  dt,
  max_reached,
  noise_mechanism = "add",
  noise_func = NULL
)
}
\description{
Simulate evidence accumulation in a drift-diffusion model
}
\keyword{internal}
