% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by.R
\name{apply_summarise_by_spec}
\alias{apply_summarise_by_spec}
\title{Internal function to apply a spec to data}
\usage{
apply_summarise_by_spec(spec_list, .data)
}
\arguments{
\item{spec_list}{A list of spec operations (the internal spec list)}

\item{.data}{A data frame}
}
\value{
A data frame with class "eam_summarise_by_tbl"
}
\description{
Internal function to apply a spec to data
}
\keyword{internal}
