% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_resample.R
\name{extract_abc_param_values}
\alias{extract_abc_param_values}
\title{Extract parameter values from abc result}
\usage{
extract_abc_param_values(abc_result)
}
\arguments{
\item{abc_result}{Single abc result object}
}
\value{
Matrix of parameter values
}
\description{
Extract parameter values from abc result
}
\keyword{internal}
