% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_chunk_result_flatten.R
\name{fill_data_table}
\alias{fill_data_table}
\title{Fill pre-allocated data.table with simulation results}
\usage{
fill_data_table(
  sim_results,
  dt_lists,
  trial_col_names,
  cond_param_names,
  first_trial_col
)
}
\arguments{
\item{sim_results}{The output from run_simulation(), a list of conditions}

\item{dt_lists}{Named list of pre-allocated vectors for each column}

\item{trial_col_names}{Character vector of trial column names}

\item{cond_param_names}{Character vector of condition parameter names}

\item{first_trial_col}{Name of the first trial column to use for item
counting}
}
\value{
Named list of filled vectors ready for data.table creation
}
\description{
This function fills the pre-allocated data.table vectors with data from
simulation results, iterating through all conditions and trials.
}
\keyword{internal}
