% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_router.R
\name{get_config_env_names}
\alias{get_config_env_names}
\title{Extract all left-hand side variable names from config formulas and prior_params}
\usage{
get_config_env_names(config)
}
\arguments{
\item{config}{A list containing simulation configuration parameters}
}
\value{
A character vector of all LHS variable names from formulas and prior_params columns
}
\description{
Extract all left-hand side variable names from config formulas and prior_params
}
\keyword{internal}
