% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_output.R
\name{map_by_condition.process_chunk}
\alias{map_by_condition.process_chunk}
\title{Process a single chunk for map_by_condition}
\usage{
map_by_condition.process_chunk(open_dataset_fn, .f, ...)
}
\arguments{
\item{open_dataset_fn}{Arrow dataset object or function that returns a dataset}

\item{.f}{Function to apply to each condition's data}

\item{...}{Additional arguments passed to .f}
}
\value{
Function that processes a chunk_idx
}
\description{
Process a single chunk for map_by_condition
}
\keyword{internal}
