% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accuracy.R
\name{plot_accuracy_graph}
\alias{plot_accuracy_graph}
\title{Plot accuracy graph (internal)}
\usage{
plot_accuracy_graph(
  accuracy_df,
  x = "item_idx",
  y = "accuracy",
  facet_x = c(),
  facet_y = c()
)
}
\arguments{
\item{accuracy_df}{Data frame with accuracy values}

\item{x}{Variable for x-axis}

\item{y}{Variable for y-axis (default: "accuracy")}

\item{facet_x}{Variables for faceting columns}

\item{facet_y}{Variables for faceting rows}
}
\value{
A ggplot2 object
}
\description{
Plot accuracy graph (internal)
}
\keyword{internal}
