% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{resolve_symbol}
\alias{resolve_symbol}
\title{Helper to resolved defined symbols in our formulas}
\usage{
resolve_symbol(expr, env, n)
}
\arguments{
\item{expr}{An expression to evaluate}

\item{env}{An environment to evaluate the expression in}

\item{n}{The number of values to generate if the expression is a distribution}
}
\value{
The evaluated value as it is, no assumption on its type
}
\description{
This function evaluates an expression in a given environment.
}
\keyword{internal}
