% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by.R
\name{summarise_by_impl}
\alias{summarise_by_impl}
\title{Internal function to perform the core summarise_by logic}
\usage{
summarise_by_impl(.data, dots, .by, .wider_by)
}
\arguments{
\item{.data}{A data frame to summarise}

\item{dots}{Quosures containing the summary expressions}

\item{.by}{Character vector of column names to group by}

\item{.wider_by}{Character vector of column names to keep as identifying columns}
}
\value{
A data frame with class "eam_summarise_by_tbl"
}
\description{
Internal function to perform the core summarise_by logic
}
\keyword{internal}
