% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accumulate_evidence_ddm_2b}
\alias{accumulate_evidence_ddm_2b}
\title{Simulate evidence accumulation in a two-bound drift-diffusion model}
\usage{
accumulate_evidence_ddm_2b(
  A_upper,
  A_lower,
  V,
  Z,
  ndt,
  max_t,
  dt,
  max_reached,
  noise_mechanism = "add",
  noise_func = NULL
)
}
\description{
Simulate evidence accumulation in a two-bound drift-diffusion model
}
\keyword{internal}
