% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accumulate_evidence_lca_gi}
\alias{accumulate_evidence_lca_gi}
\title{Simulate evidence accumulation in a leaky competing accumulator model with
Global Inhibition (LCA-GI)}
\usage{
accumulate_evidence_lca_gi(
  A,
  V,
  Z,
  ndt,
  beta,
  k,
  max_t,
  dt,
  max_reached,
  noise_func = NULL
)
}
\description{
Simulate evidence accumulation in a leaky competing accumulator model with
Global Inhibition (LCA-GI)
}
\keyword{internal}
