% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_chunk_result_flatten.R
\name{calculate_total_rows}
\alias{calculate_total_rows}
\title{Calculate total number of rows needed for flattened data}
\usage{
calculate_total_rows(sim_results, first_trial_col)
}
\arguments{
\item{sim_results}{The output from run_simulation(), a list of conditions}

\item{first_trial_col}{Name of the first trial column to use for counting}
}
\value{
Integer, total number of rows needed
}
\description{
This function counts the total number of items across all conditions and
trials to determine the size needed for pre-allocation.
}
\keyword{internal}
