% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{evaluate_with_dt}
\alias{evaluate_with_dt}
\title{Evaluate a list of formulas sequentially with data}
\usage{
evaluate_with_dt(formulas, data = list(), n)
}
\arguments{
\item{formulas}{A list of formulas to evaluate}

\item{data}{A list of named values to use as the initial environment}

\item{n}{The number of values to generate for each formula}
}
\value{
A named list of evaluated values with length n
}
\description{
This function evaluates a list of formulas sequentially, allowing later
formulas to reference
}
\keyword{internal}
