% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_router.R
\name{get_backend_detectors}
\alias{get_backend_detectors}
\title{Get all registered backend detectors}
\usage{
get_backend_detectors()
}
\value{
A list of backend detector functions
}
\description{
Get all registered backend detectors
}
\keyword{internal}
