% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_output.R
\name{map_by_condition.parallel.heuristic}
\alias{map_by_condition.parallel.heuristic}
\title{Heuristic to determine if parallel processing should be used}
\usage{
map_by_condition.parallel.heuristic(chunk_indices)
}
\arguments{
\item{chunk_indices}{Vector of chunk indices}
}
\value{
Logical value indicating whether to use parallel processing
}
\description{
Heuristic to determine if parallel processing should be used
}
\keyword{internal}
