% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_output.R
\name{new_simulation_output}
\alias{new_simulation_output}
\title{Create a eam_simulation_output object}
\usage{
new_simulation_output(simulation_config, output_dir)
}
\description{
Create a eam_simulation_output object
}
\keyword{internal}
