% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accuracy.R
\name{plot_accuracy_ddm_2b}
\alias{plot_accuracy_ddm_2b}
\title{Plot accuracy for DDM-2B model (internal)}
\usage{
plot_accuracy_ddm_2b(
  simulated_output,
  observed_df,
  x = "item_idx",
  facet_x = c(),
  facet_y = c()
)
}
\arguments{
\item{simulated_output}{Simulation output object}

\item{observed_df}{Observed data frame}

\item{x}{Variable for x-axis}

\item{facet_x}{Variables for faceting columns}

\item{facet_y}{Variables for faceting rows}
}
\value{
A ggplot2 object
}
\description{
Plot accuracy for DDM-2B model (internal)
}
\keyword{internal}
