% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by.R
\name{+.eam_summarise_by_spec}
\alias{+.eam_summarise_by_spec}
\title{Add two summarise_by specs together}
\usage{
\method{+}{eam_summarise_by_spec}(e1, e2)
}
\arguments{
\item{e1}{First eam_summarise_by_spec or eam_summarise_by_tbl object}

\item{e2}{Second eam_summarise_by_spec or eam_summarise_by_tbl object}
}
\value{
A combined eam_summarise_by_spec object
}
\description{
S3 method for the + operator to combine two `eam_summarise_by_spec`
objects into a single spec that will apply both operations.
}
