% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by.R
\name{+.eam_summarise_by_tbl}
\alias{+.eam_summarise_by_tbl}
\title{Join two eam_summarise_by_tbl objects}
\usage{
\method{+}{eam_summarise_by_tbl}(e1, e2)
}
\arguments{
\item{e1}{First eam_summarise_by_tbl object}

\item{e2}{Second eam_summarise_by_tbl object}
}
\value{
A joined data frame with class "eam_summarise_by_tbl",
  preserving the .wider_by attribute from the input tables
}
\description{
S3 method for the + operator to join two summary tables created by
\code{summarise_by}. Tables must have identical .wider_by attributes
to be joined.
}
